requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
mysqlclient==2.1.0
Pillow==9.1.0
sqlparse==0.4.2
